Adds the ability to format submenu items as button
If "button" attribute is set to True on submenu item, format it as a single "button". Supported attributes are copied from the "button" component. Otherwise, preserve previous behavior. Icon can still be added independently, if defined.
Define "compact"=True on submenu, to remove extra spaing on the right and make sure that icon-only button are stacked (placed one per line).