Added create-migration command line instruction (#943)
* Added create-migration command line instruction

This update adds a command line interface to create migrations directly with sqlpage, and have it manage the timestamps and uniqueness of migration names.

* Update sqlpage/migrations/README.md

Co-authored-by: Ophir LOJKINE <contact@ophir.dev>

* Update sqlpage/migrations/README.md

Co-authored-by: Ophir LOJKINE <contact@ophir.dev>

* Updated with changes for config directory, migration output

Now respects the configuration_directory environment variable.

It also outputs the path of the new migration created when it is created, and takes care to display that path relative to the current working directory.

Lastly, the execution of create-migration command was moved above the rest of the initialization, so that creating a new migration does not run existing migrations. This allows you to create multiple migrations before you run sqlpage normally again to execute them.

---------

Co-authored-by: Ophir LOJKINE <contact@ophir.dev>